pyproject.toml
[tool.poetry]
name = "coapartybot"
version = "0.1.0"
description = ""
authors = ["PotatoMind"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
"discord.py" = "^2.0.1"
redis = "^4.3.4"
numpy = "^1.23.2"
DiscordUtils = "^1.3.4"
textdistance = "^4.4.0"
bs4 = "^0.0.1"
motor = "^3.0.0"
matplotlib = "^3.5.3"
psutil = "^5.9.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
